fix warnings about module names
